log: move platform specific handlers to oslayer
