+++ plover/gui/log.py
@@ -6,10 +6,10 @@ handler = None
 
 try:
     if sys.platform.startswith('linux'):
-        from plover.gui.log_dbus import DbusNotificationHandler
+        from plover.oslayer.log_dbus import DbusNotificationHandler
         handler = DbusNotificationHandler
     elif sys.platform.startswith('darwin'):
-        from plover.gui.log_osx import OSXNotificationHandler
+        from plover.oslayer.log_osx import OSXNotificationHandler
         handler = OSXNotificationHandler
 except Exception as e:
     log.info('could not import platform gui log', exc_info=e)
